Segment _TEXT exceeds 64K error removed
The error "Segment _TEXT exceeds 64K" is removed after a bit of refactoring, but a new linker error has cropped up: "Undefined symbol hospital::get_med_by_code(int) in module IFACE.CPP"
The program works on the 'Large' memory model or the 'Huge' memory model. Anything lower than that and the 64K error crops up again.
The 2 .EXT files are the dump files of hosp.cpp and iface.cpp created using tdump.exe.
